moved CarouselLayout.js up a folder. This is because CarouselLayout.js is now being sued by multiple files and no longer belongs in the browse folder.
