+++ mediabox/src/main/resources/static/css/account.css
@@ -42,24 +42,18 @@ input {
     width: 100%
 }
 
-input[type=button], input[type=submit], input[type=reset]{
-    color: #f9f9f9;
-    border: 1px solid #ccc;
-    border-radius: 4px;
-    padding: 4px 8px;
-    text-decoration: none;
-    background-color: #5e9bff;
-    cursor: pointer;
-    width: 100%;
-}
+
 
 input[type=submit]{
-    position: relative;
-    top: 5em;
+    width: 50%;
 }
 
 .lowerDiv {
     position: relative;
     top: 10em;
     width: 100%;
+}
+
+#hide {
+    visibility: hidden;
 }
